feat(csa): add code health agent skill
Add the bundled code-health-agent executor skill with non-blocking health scans, token-threshold configuration, SRP checks, Code Judo analysis, and GitHub issue proposal workflow.\n\nExtend skill resolution to find bundled CSA namespace skills by name, and split skill_resolver tests out of the production file to keep the edited module under the 8K token health gate.\n\nBump workspace version to 0.1.790 for the feature branch pre-commit gate.
