+++ skills/csa/code-health-agent/.skill.toml
@@ -1,0 +1,10 @@
+[skill]
+name = "code-health-agent"
+version = "0.1.0"
+
+[agent]
+tier = "tier-2-standard"
+max_turns = 20
+token_budget = 120000
+workspace_access = "read-only"
+tools = [{ tool = "auto" }]
